test: Cleanup GPU-AV BufferDeviceAddress test
