+++ tests/unit/gpu_av_positive.cpp
@@ -213,85 +213,6 @@ TEST_F(PositiveGpuAV, SetSSBOPushDescriptor) {
     vk::DestroyPipeline(m_device->device(), pipeline, nullptr);
 }
 
-TEST_F(PositiveGpuAV, BufferDeviceAddress) {
-    TEST_DESCRIPTION("Makes sure that writing to a buffer that was created after command buffer record doesn't get OOB error");
-    SetTargetApiVersion(VK_API_VERSION_1_2);
-    AddRequiredExtensions(VK_KHR_BUFFER_DEVICE_ADDRESS_EXTENSION_NAME);
-    RETURN_IF_SKIP(InitGpuAvFramework());
-    VkPhysicalDeviceBufferDeviceAddressFeaturesKHR bda_features = vku::InitStructHelper();
-    VkPhysicalDeviceFeatures2KHR features2 = GetPhysicalDeviceFeatures2(bda_features);
-    if (!features2.features.shaderInt64) {
-        GTEST_SKIP() << "shaderInt64 is not supported";
-    }
-    features2.features.robustBufferAccess = VK_FALSE;
-
-    RETURN_IF_SKIP(InitState(nullptr, &features2));
-    InitRenderTarget();
-
-    // Make a uniform buffer to be passed to the shader that contains the pointer and write count
-    uint32_t buffer_size = 12;  // 64 bit pointer + int
-    VkMemoryPropertyFlags mem_props = VK_MEMORY_PROPERTY_HOST_VISIBLE_BIT | VK_MEMORY_PROPERTY_HOST_COHERENT_BIT;
-    vkt::Buffer buffer0(*m_device, buffer_size, VK_BUFFER_USAGE_UNIFORM_BUFFER_BIT, mem_props);
-
-    char const *shader_source = R"glsl(
-        #version 450
-        #extension GL_EXT_buffer_reference : enable
-        layout(buffer_reference, buffer_reference_align = 16) buffer bufStruct;
-        layout(set = 0, binding = 0) uniform ufoo {
-            bufStruct data;
-            int nWrites;
-        } u_info;
-        layout(buffer_reference, std140) buffer bufStruct {
-            int a[4];
-        };
-        void main() {
-            for (int i=0; i < u_info.nWrites; ++i) {
-                u_info.data.a[i] = 0xdeadca71;
-            }
-        }
-    )glsl";
-    VkShaderObj vs(this, shader_source, VK_SHADER_STAGE_VERTEX_BIT);
-
-    CreatePipelineHelper pipe(*this);
-    pipe.InitState();
-    pipe.shader_stages_ = {vs.GetStageCreateInfo()};
-    pipe.rs_state_ci_.rasterizerDiscardEnable = VK_TRUE;
-    pipe.dsl_bindings_ = {{0, VK_DESCRIPTOR_TYPE_UNIFORM_BUFFER, 1, VK_SHADER_STAGE_ALL, nullptr}};
-    pipe.CreateGraphicsPipeline();
-
-    pipe.descriptor_set_->WriteDescriptorBufferInfo(0, buffer0.handle(), 0, sizeof(uint32_t));
-    pipe.descriptor_set_->UpdateDescriptorSets();
-
-    m_commandBuffer->begin();
-    m_commandBuffer->BeginRenderPass(m_renderPassBeginInfo);
-    vk::CmdBindPipeline(m_commandBuffer->handle(), VK_PIPELINE_BIND_POINT_GRAPHICS, pipe.Handle());
-    vk::CmdBindDescriptorSets(m_commandBuffer->handle(), VK_PIPELINE_BIND_POINT_GRAPHICS, pipe.pipeline_layout_.handle(), 0, 1,
-                              &pipe.descriptor_set_->set_, 0, nullptr);
-    vk::CmdDraw(m_commandBuffer->handle(), 3, 1, 0, 0);
-    m_commandBuffer->EndRenderPass();
-    m_commandBuffer->end();
-
-    // Make another buffer to write to
-    buffer_size = 64;  // Buffer should be 16*4 = 64 bytes
-    VkMemoryAllocateFlagsInfo allocate_flag_info = vku::InitStructHelper();
-    allocate_flag_info.flags = VK_MEMORY_ALLOCATE_DEVICE_ADDRESS_BIT;
-    vkt::Buffer buffer1(*m_device, buffer_size, VK_BUFFER_USAGE_SHADER_DEVICE_ADDRESS_BIT_KHR, mem_props, &allocate_flag_info);
-
-    // Get device address of buffer to write to
-    auto pBuffer = buffer1.address();
-
-    auto *data = static_cast<VkDeviceAddress *>(buffer0.memory().map());
-    data[0] = pBuffer;
-    data[1] = 4;
-    buffer0.memory().unmap();
-
-    VkSubmitInfo submit_info = vku::InitStructHelper();
-    submit_info.commandBufferCount = 1;
-    submit_info.pCommandBuffers = &m_commandBuffer->handle();
-    vk::QueueSubmit(m_default_queue, 1, &submit_info, VK_NULL_HANDLE);
-    vk::QueueWaitIdle(m_default_queue);
-}
-
 // Regression test for semaphore timeout with GPU-AV enabled:
 // https://github.com/KhronosGroup/Vulkan-ValidationLayers/issues/4968
 TEST_F(PositiveGpuAV, GetCounterFromSignaledSemaphoreAfterSubmit) {
